Spider component guide, run vr against previews
We can’t easily get a list of URLs for a component guide. Instead use
the wraith spider feature and crawl the review domain’s component
guide. Exclude paths that don’t start with /component-guide and only
compare paths that end with `/preview` to limit the number of pages and
speed up the test run.

* Include rake tasks for running locally
* Use Heroku runner in app